+++ jbbp/src/main/java/com/igormaznitsa/jbbp/model/JBBPFieldBit.java
@@ -74,7 +74,7 @@ public final class JBBPFieldBit extends JBBPAbstractField implements JBBPNumeric
 
   @Override
   public double getAsDouble() {
-    return (double) (this.value & 0xFF);
+    return this.value & 0xFF;
   }
 
   @Override
