+++ jbbp/src/main/java/com/igormaznitsa/jbbp/model/JBBPFieldShort.java
@@ -52,17 +52,17 @@ public final class JBBPFieldShort extends JBBPAbstractField implements JBBPNumer
     final int b0 = JBBPUtils.reverseBitsInByte((byte) value) & 0xFF;
     final int b1 = JBBPUtils.reverseBitsInByte((byte) (value >> 8)) & 0xFF;
 
-    return (long) ((short) (b0 << 8) | (short) b1);
+    return (short) (b0 << 8) | (short) b1;
   }
 
   @Override
   public double getAsDouble() {
-    return (double) this.value;
+    return this.value;
   }
 
   @Override
   public float getAsFloat() {
-    return (float) this.value;
+    return this.value;
   }
 
   @Override
